refactor(devframe)!: keep utils/hash; move hub-only internals to devframe/internal
- Restore `devframe/utils/hash` — kept as a public utility (maintainer request).
- Add a new `devframe/internal` entry point for the low-level primitives shared
  only between `devframe` and `@devframes/hub`, and relocate `createContextRpcServer`,
  `DevframeAgentHost`, and `coerceAgentPositionalArgs` there from `devframe/node`.
  `devframe/internal` is an explicitly-unstable cross-package surface; the hub
  imports these three from it. This keeps the public `devframe/node` barrel to
  the genuinely-public server-assembly API.

Updates the migration guide, SKILL/utilities docs, knip entry list, plan, and
regenerates the tsnapi snapshots.
