+++ packages/devframe/package.json
@@ -29,6 +29,7 @@
     "./constants": "./dist/constants.mjs",
     "./helpers/vite": "./dist/helpers/vite.mjs",
     "./initiate": "./dist/adapters/initiate.mjs",
+    "./internal": "./dist/internal/index.mjs",
     "./node": "./dist/node/index.mjs",
     "./node/auth": "./dist/node/auth.mjs",
     "./node/hub-internals": "./dist/node/hub-internals.mjs",
@@ -46,6 +47,7 @@
     "./utils/colors": "./dist/utils/colors.mjs",
     "./utils/crypto-token": "./dist/utils/crypto-token.mjs",
     "./utils/events": "./dist/utils/events.mjs",
+    "./utils/hash": "./dist/utils/hash.mjs",
     "./utils/launch-editor": "./dist/utils/launch-editor.mjs",
     "./utils/nanoid": "./dist/utils/nanoid.mjs",
     "./utils/open": "./dist/utils/open.mjs",
