+++ packages/devframe/src/node/index.ts
@@ -1,20 +1,17 @@
 // Node-side public API for hosts that wire up their own runtime — the
 // server-assembly surface (`createHostContext` → `startHttpAndWs`), the
-// transport-agnostic RPC core (`createContextRpcServer`), the instance
-// registry, storage, and the two URL/host helpers consumers use.
+// instance registry, storage, and the two URL/host helpers consumers use.
 //
 // The diagnostics/services/views hosts, the streaming/shared-state/scope/
 // settings factories, and the internal host helpers (`toDialableHost`,
 // `formatHostForUrl`) stay internal. `toAgentToolName` lives at
 // `devframe/utils/agent-tool-name` (a client-safe string transform).
 //
-// `coerceAgentPositionalArgs`, the `DevframeAgentHost` class, and
-// `createContextRpcServer` stay public because `@devframes/hub` composes its
-// own commands/agent host and hand-rolled server on top of them — the exact
-// custom-transport path `createContextRpcServer` exists for.
-export { coerceAgentPositionalArgs } from './agent-args'
+// The low-level primitives shared only between `devframe` and `@devframes/hub`
+// — `createContextRpcServer`, `DevframeAgentHost`, `coerceAgentPositionalArgs`
+// — live at `devframe/internal` (an explicitly-unstable cross-package surface),
+// not here.
 export * from './context'
-export { DevframeAgentHost } from './host-agent'
 // `RpcFunctionsHostImpl` stays internal; expose only the structural
 // `RpcFunctionsHost` type so consumers can type/cast `ctx.rpc` without
 // pulling in the implementation's `@internal` members.
@@ -26,10 +23,6 @@ export * from './host-h3'
 // lower-level read/probe/prune helpers stay internal to the connector.
 export { listLiveDevframeInstances, registerDevframeInstance } from './instance-registry'
 export type { DevframeInstanceRecord, DevframeInstanceRegistration } from './instance-registry'
-// The transport-agnostic RPC core is public so hosts that bind their own
-// transports (a Bun fetch-upgrade route, a custom relay — e.g. `initHub`)
-// reuse the exact session/auth wiring `startHttpAndWs` uses.
-export * from './rpc-core'
 export * from './server'
 export * from './storage'
 export { isObject, normalizeHttpServerUrl } from './utils'
