+++ packages/devframe/tsdown.config.ts
@@ -73,6 +73,7 @@ const clientEntries = {
   'utils/colors': 'src/utils/colors.ts',
   'utils/crypto-token': 'src/utils/crypto-token.ts',
   'utils/events': 'src/utils/events.ts',
+  'utils/hash': 'src/utils/hash.ts',
   'utils/nanoid': 'src/utils/nanoid.ts',
   'utils/simple-schema': 'src/utils/simple-schema.ts',
   'utils/shared-state': 'src/utils/shared-state.ts',
@@ -96,6 +97,7 @@ const serverEntries = {
   'node/index': 'src/node/index.ts',
   'node/auth': 'src/node/auth/index.ts',
   'node/hub-internals': 'src/node/hub-internals/index.ts',
+  'internal/index': 'src/internal/index.ts',
   'utils/launch-editor': 'src/utils/launch-editor.ts',
   'utils/open': 'src/utils/open.ts',
   'utils/serve-static': 'src/utils/serve-static.ts',
@@ -146,6 +148,7 @@ export default defineConfig([
             resolve(distDir, 'utils/colors.mjs'),
             resolve(distDir, 'utils/crypto-token.mjs'),
             resolve(distDir, 'utils/events.mjs'),
+            resolve(distDir, 'utils/hash.mjs'),
             resolve(distDir, 'utils/nanoid.mjs'),
             resolve(distDir, 'utils/simple-schema.mjs'),
             resolve(distDir, 'utils/shared-state.mjs'),
