+++ plans/032-0.9-public-api-surface-reduction.md
@@ -90,10 +90,17 @@ consumer):
   hand-rolled-server host the ⚠️ note below anticipated. **Kept** (resolves the
   open question: keep it).
 
-Net effect: dead-subpath removals are `utils/hash`, `utils/promise`, and
-`utils/scope` (not `ws-bun`); the `devframe/node` trim drops **9** internal
-exports, not 12. `devframe/adapters/embedded` was initially removed but restored
-on maintainer request — it stays as a named, discoverable adapter.
+Net effect (after maintainer follow-ups):
+
+- Dead-subpath removals are `utils/promise` and `utils/scope` only. `ws-bun` is
+  kept (hub uses it); `utils/hash` is kept (maintainer request); `embedded` is
+  kept (maintainer request — a named, discoverable adapter).
+- The `devframe/node` trim drops **12** exports again — but the three
+  hub-only low-level APIs (`createContextRpcServer`, `DevframeAgentHost`,
+  `coerceAgentPositionalArgs`) are **relocated** to a new **`devframe/internal`**
+  entry (an explicitly-unstable cross-package surface) rather than kept on the
+  public `devframe/node`, and the hub imports them from there. The other 9 stay
+  fully internal (relative imports only).
 
 ### ⚠️ One item to confirm during review
